# perf-simple-query: add cpu_cycles / op metric