--- a/test/perf/perf.hh
+++ b/test/perf/perf.hh
@@ -54,6 +54,7 @@ struct executor_shard_stats {
     uint64_t log_allocations = 0;
     uint64_t tasks_executed = 0;
     uint64_t instructions_retired = 0;
+    uint64_t cpu_cycles_retired = 0;
     uint64_t errors = 0;
 };
 
@@ -65,6 +66,7 @@ operator+(executor_shard_stats a, executor_shard_stats b) {
     a.log_allocations += b.log_allocations;
     a.tasks_executed += b.tasks_executed;
     a.instructions_retired += b.instructions_retired;
+    a.cpu_cycles_retired += b.cpu_cycles_retired;
     a.errors += b.errors;
     return a;
 }
@@ -77,6 +79,7 @@ operator-(executor_shard_stats a, executor_shard_stats b) {
     a.log_allocations -= b.log_allocations;
     a.tasks_executed -= b.tasks_executed;
     a.instructions_retired -= b.instructions_retired;
+    a.cpu_cycles_retired -= b.cpu_cycles_retired;
     a.errors -= b.errors;
     return a;
 }
@@ -97,6 +100,7 @@ class executor {
     uint64_t _count;
     uint64_t _errors;
     linux_perf_event _instructions_retired_counter = linux_perf_event::user_instructions_retired();
+    linux_perf_event _cpu_cycles_retired_counter = linux_perf_event::user_cpu_cycles_retired();
 private:
     executor_shard_stats executor_shard_stats_snapshot();
     future<> run_worker() {
@@ -127,11 +131,13 @@ public:
     future<executor_shard_stats> run() {
         auto stats_start = executor_shard_stats_snapshot();
         _instructions_retired_counter.enable();
+        _cpu_cycles_retired_counter.enable();
         auto idx = boost::irange(0, (int)_n_workers);
         return parallel_for_each(idx.begin(), idx.end(), [this] (auto idx) mutable {
             return this->run_worker();
         }).then([this, stats_start] {
             _instructions_retired_counter.disable();
+            _cpu_cycles_retired_counter.disable();
             auto stats_end = executor_shard_stats_snapshot();
             return stats_end - stats_start;
         });
@@ -151,6 +157,7 @@ executor<Func>::executor_shard_stats_snapshot() {
         .log_allocations = perf_logallocs(),
         .tasks_executed = perf_tasks_processed(),
         .instructions_retired = _instructions_retired_counter.read(),
+        .cpu_cycles_retired = _cpu_cycles_retired_counter.read(),
         .errors = _errors,
     };
 }
@@ -161,6 +168,7 @@ struct perf_result {
     double logallocs_per_op;
     double tasks_per_op;
     double instructions_per_op;
+    double cpu_cycles_per_op;
     uint64_t errors;
 };
 
@@ -232,6 +240,7 @@ std::vector<Res> time_parallel_ex(Func func, unsigned concurrency_per_core, int 
         result.logallocs_per_op = double(stats.log_allocations) / stats.invocations;
         result.tasks_per_op = double(stats.tasks_executed) / stats.invocations;
         result.instructions_per_op = double(stats.instructions_retired) / stats.invocations;
+        result.cpu_cycles_per_op = double(stats.cpu_cycles_retired) / stats.invocations;
         result.errors = stats.errors;
 
         uf(result, stats);