@bhalevy 2024-05-22
Comparing the noise levels:
| stat | avg | std | std/avg |
| -- | -- | -- | -- |
| cycles/op | 22851.5 | 216.33 | 0.0095 |
| tps | 89366.44 | 1379.14 | 0.0154 |
| insns/op | 42352.5 | 18.86 | 0.0004 |