@bhalevy 2024-05-22
> #### 🔴 CI State: FAILURE
> ❌ - [Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/9070/)
> 
> #### Build Failure:
> ```
> [2024-05-22T09:57:59.545Z] FAILED: build/debug/test/perf/perf_alternator.o 
> [2024-05-22T09:57:59.545Z] clang++ -MD -MT build/debug/test/perf/perf_alternator.o -MF build/debug/test/perf/perf_alternator.o.d -std=c++20 -I/jenkins/workspace/scylla-master/scylla-ci/scylla/seastar/include -I/jenkins/workspace/scylla-master/scylla-ci/scylla/build/debug/seastar/gen/include -U_FORTIFY_SOURCE -Werror=unused-result -fstack-clash-protection -fsanitize=address -fsanitize=undefined -fno-sanitize=vptr -DSEASTAR_API_LEVEL=7 -DSEASTAR_BUILD_SHARED_LIBS -DSEASTAR_SSTRING -DSEASTAR_LOGGER_COMPILE_TIME_FMT -DSEASTAR_SCHEDULING_GROUPS_COUNT=16 -DSEASTAR_DEBUG -DSEASTAR_DEFAULT_ALLOCATOR -DSEASTAR_SHUFFLE_TASK_QUEUE -DSEASTAR_DEBUG_SHARED_PTR -DSEASTAR_DEBUG_PROMISE -DSEASTAR_LOGGER_TYPE_STDOUT -DSEASTAR_TYPE_ERASE_MORE -DBOOST_NO_CXX98_FUNCTION_BASE -DFMT_SHARED -I/usr/include/p11-kit-1 -ffile-prefix-map=/jenkins/workspace/scylla-master/scylla-ci/scylla=. -march=westmere -DDEBUG -DSANITIZE -DDEBUG_LSA_SANITIZER -DSCYLLA_ENABLE_ERROR_INJECTION -Og -DSCYLLA_BUILD_MODE=debug -g -gz -iquote. -iquote build/debug/gen -std=gnu++20  -ffile-prefix-map=/jenkins/workspace/scylla-master/scylla-ci/scylla=. -march=westmere -DBOOST_ALL_DYN_LINK    -fvisibility=hidden -isystem abseil -Wall -Werror -Wextra -Wimplicit-fallthrough -Wno-mismatched-tags -Wno-c++11-narrowing -Wno-overloaded-virtual -Wno-unused-parameter -Wno-unsupported-friend -Wno-missing-field-initializers -Wno-deprecated-copy -Wno-psabi -Wno-enum-constexpr-conversion -Wno-error=deprecated-declarations -DXXH_PRIVATE_API -DSEASTAR_TESTING_MAIN  -c -o build/debug/test/perf/perf_alternator.o test/perf/perf_alternator.cc
> [2024-05-22T09:57:59.555Z] In file included from test/perf/perf_alternator.cc:24:
> [2024-05-22T09:57:59.555Z] ./test/perf/perf.hh:103:70: error: no member named user_cpu_cycles_retired in linux_perf_event
> [2024-05-22T09:57:59.555Z]     linux_perf_event _cpu_cycles_retired_counter = linux_perf_event::user_cpu_cycles_retired();
> [2024-05-22T09:57:59.555Z]                                                    ~~~~~~~~~~~~~~~~~~^
> [2024-05-22T09:57:59.555Z] 1 error generated.
> ```
> 
> #### Build Details:
> * Duration: 37 min
> * Builder: i-06a6117050b9e5772 (m5ad.8xlarge)

Sorry, there's a dependency on a seastar change that I need to submit as well.