@bhalevy 2024-05-22
Draft until https://github.com/scylladb/seastar/pull/2261 gets to the scylladb repo