@bhalevy 2024-06-14
@scylladb/scylla-maint please merge